+++ src/components/UserPanel/userSlice.ts
@@ -1,0 +1,17 @@
+import { createSlice } from '@reduxjs/toolkit';
+
+export const slice = createSlice({
+  name: 'user',
+  initialState: {
+    displayName: 'No custom name',
+  },
+  reducers: {
+    changeDisplayName: (state, action) => {
+      state.displayName = action.payload.newDisplayName
+    }
+  }
+});
+
+export const selectDisplayName = (state: any) : string => state.user.displayName;
+export const { changeDisplayName } = slice.actions
+export default slice.reducer;