+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog for Stock Dashboard
 
+## 2020-03-09 - 0.2.0
+
+- Add Redux simple counter with config
+- Add test engine AVA
+- MirageJS for local backend mocking
+- Init components for Stock Table, UserPanel and bottom navigation
 
 ## 2020-01-15 - 0.1.0